Use hardware volume keys from anywhere inside Kore
* Move dispatchKeyEvent callback for volume control to the BaseMediaActivity to allow all activities with now playing panel to receive hardware button events
* Pressing the hardware buttons expand the now playing panel for 2 seconds
** If there is button clicked in the now playing panel it'll stay expanded

resolves xbmc/Kore#235
